Renamed src/utils/date-format.js module to src/utils/date.js.
